pyproject.toml
[tool.poetry]
name = "samsul"
version = "0.1.0"
description = ""
authors = ["ken.kustian <ken.kustian@gmail.com>"]
readme = "README.md"
packages = [{include = "samsul", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.8"
authlib = "^1.3.1"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
python-ulid = "^2.7.0"
pypdf = "^4.2.0"
streaming-form-data = "^1.15.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
